+++ NEWS.md
@@ -9,7 +9,8 @@
 * Add `fit$internal$packageVersion` to allow `predict`, `cAIC` etc to check
   and throw error if there's a package inconsistency between object and installed
   package
-* Add `simulate_conditional_gmrf` as precursor to efficient projections
+* Add `simulate_conditional_gmrf` to do conditional simulations from a GMRF
+* Add `project` to project tinyVAST forward in time
 
 # tinyVAST 1.1.1